Explored using Intake data
Pulled intake data from austin.gov site and attempted to match it to
the training/test data.  The train data can be matched on AnimalID;
however, the test data does not have AnimalID. As such, very minimal
amounts of matches can be made to the test data rendering the data semi
useless for prediction.